refactor: move inputs definition and types into their own file
